Refactore Header component
