+++ src/components/Dragons/Dragons.jsx
@@ -1,0 +1,7 @@
+const Dragons = () => {
+  return(
+    <></>
+  )
+}
+
+export default Dragons;
