+++ src/components/Header/NavBar.css
@@ -22,7 +22,7 @@
   gap: 1rem;
 }
 
-.nav-items li:nth-of-type(2) {
+.nav-items li:nth-of-type(3) {
   border-right: 1.5px solid black;
   padding-right: 1rem;
 }
